+++ bulldog.core/src/main/java/org/bulldog/core/gpio/base/AbstractPwm.java
@@ -8,7 +8,7 @@ public abstract class AbstractPwm extends AbstractPinFeature implements Pwm {
 	private String NAME_FORMAT = "PWM - status '%s' - frequency '%d' with duty '%.2f' on Pin %s";
 
 	private float duty = 0.0f;
-	private long frequency = 0;
+	private float frequency = 0;
 	private boolean enabled = false;
 	
 	public AbstractPwm(Pin pin) {
@@ -49,12 +49,12 @@ public abstract class AbstractPwm extends AbstractPinFeature implements Pwm {
 		return duty;
 	}
 	
-	public void setFrequency(long frequency) {
+	public void setFrequency(float frequency) {
 		this.frequency = frequency;
 		setPwmImpl(frequency, getDuty());
 	}
 	
-	public long getFrequency() {
+	public float getFrequency() {
 		return frequency;
 	}
 
@@ -62,7 +62,7 @@ public abstract class AbstractPwm extends AbstractPinFeature implements Pwm {
 		return false;
 	}
 	
-	protected abstract void setPwmImpl(long frequency, float duty);
+	protected abstract void setPwmImpl(float frequency, float duty);
 	protected abstract void enableImpl();
 	protected abstract void disasbleImpl();
